requirements.txt
python-telegram-bot==20.7
websockets==12.0
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
numpy>=1.21.0
pandas>=1.3.0
ta-lib>=0.4.0
MetaTrader5>=5.0.45